+++ components/index.ts
@@ -14,6 +14,7 @@ export { default as Footer } from './Footer'
 export { default as CandyMachineList } from './CandyMachineList'
 export { default as NetworkSelector } from './NetworkSelector'
 export { default as NetworkTps } from './NetworkTps'
+export { default as Breadcrumb } from './Breadcrumb'
 
 export * from './cards'
 export * from './forms'
